Huge improvements to the codebase.
I've switched the data structure of code representation to a proper tree, allowing for an accurate cyclomatic complexity algorithm and for much simpler parsing and inheritance/scope detection, and eventually parsing methods and functions if required.

I've hugely improved comment detection and line counting, making it so that docstrings can now be properly included in the number of lines of comments.

'JavaSourceFileParser' was bloated and hard to read, so I've gutted it and gave its functionality to its subclasses. Especially now that we have a tree, parsing no longer needs to be linear, but rather it is now dynamically programmed through recursive node calls in 'ParsedClass' and 'CodeTree' constructors.

I've also allocated all regex calls to the StringParser class and just in general made things more modular. Huge quality of life improvement for myself with just that.